finish track view ui
