+++ lib/features/home/presentation/views/widgets/tracks_list_view.dart
@@ -1,0 +1,28 @@
+import 'package:final_project_iti/features/home/presentation/views/widgets/track_item.dart';
+import 'package:flutter/material.dart';
+
+import '../../../data/models/track_model.dart';
+
+class TracksListView extends StatelessWidget {
+  const TracksListView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final List<TrackModel> trackItems = TrackModel.tracks;
+    return ListView.builder(
+      physics: const BouncingScrollPhysics(),
+      itemCount: trackItems.length,
+      itemBuilder: (context, index) {
+        final track = trackItems[index];
+        return Padding(
+          padding: const EdgeInsets.only(bottom: 16.0),
+          child: TrackItem(
+            title: track.title,
+            description: track.description,
+            imageUrl: track.image,
+          ),
+        );
+      },
+    );
+  }
+}
